[bt] Update for rustc write-only warnings.
New rust versions have a warning on a struct field that is written but
never read.  Update the fields used for holding resources, and remove
fields that are not used, to be readded later if needed.

Fixed: 69489
Test: no behavior changes; fx test
Change-Id: I502af731153110c293ea3e62f6f7f711bc1e2cd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483799
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>