[make-fuchsia-vol] Big refactor to improve testability.
This CL splits "main" into several functions: initArguments() for
parsing arguments, prepareDisk() for resizing and determining
information about the disk, createPartitionTable() for figuring out the
sizes of the partitions and the structure of the GPT, and writeDisk() for
writing out the disk image.

The intent behind this refactor is to improve the testability of
make-fuchsia-vol, particularly createPartitionTable(). For simplicity,
the majority of functions are minimally altered (they still reference
the global argument variables), but createPartitionTable() is modified
to be hermetic and not rely on the global variables. It takes all
configuration as arguments to the function.

No functional change is expected from this CL - it is a pure refactor
that adds tests for createPartitionTable().

Change-Id: I51af963021a9b963cc7a396e7fd1e84f7f4ce77e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481823
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>