[blobfs] Fix unlink behaviour while clones exist
Blob::Purgeable was missing a check that there are no active VMO clones.
Unlinking the blob when there are no active VMO clones is unsafe since
the pager is holding onto a reference to the node index for the blob
(which would be unallocated by the unlink).

Also made AllocatedExtentIterator defensively check if the inode is
allocated, which made testing easier (otherwise the failure would be
silent, as the old inode was not overwritten but simply marked
unallocated, so still had good data). This required a small change in
Blobfs::FreeInode, which was creating an allocator *after* freeing the
inode (although before persisting that change).

Added a new unit test which reliably triggers the issue.

Bug: b/177017725
Multiply: blobfs-unittest
Change-Id: I7485da70eb6790326201ebf3963668263a0d4a04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483499
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
(cherry picked from commit b76ceb09dcde2e689e65355259364220b3314872)