[lowpan][lowpan-spinel-driver] Use empty identity when not provisioned
Previously, we were filling out the identity structure for the
lowpan network even if we were not provisioned. This change makes
sure that the identity is empty unless the interface is actually
provisioned with a network.

This change also updates `lowpanctl` to make sure that it prints out
the content of the identity structure even when not provisioned. 

Bug: 69572
Test: fx test lowpan-integration-tests
Change-Id: Ie8aabb8c8f114e7d227405dc3338b3a4995f2529
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484458
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Fuchsia-Auto-Submit: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>