[system-update-committer] do commit when failure ignored
This change refactors the system-update-committer to make its behavior
consistent with the FIDL documentation for IsCurrentSystemCommitted.

Specifically, we actually complete the commit if a verification failure
is supposed to be ignored. This is different from the previous behavior,
which did not complete the commit upon an ignored failed check.

Test: system-update-committer-tests
Test: system-update-committer-integration-tests
Fixed: 69537
Change-Id: Iad1cd451c016c876235b890500f0bfcf252147a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483138
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>