(Attempt to) add an initial volume
mkfs adds a volume named "default", which mount will open up.

Adds a "root object ID" to object stores.

Volumes are managed in a directory which is the root object for the root
object store. Entries in the directory point to object stores which root
each volume; the root object ID for these volume object stores is the
root directory.

Added a test which unfortunately fails ATM (disabled_test_add_volume).
The volume doesn't appear after remounting the filesystem.

Change-Id: I45c6f2dc5ae0fb1644cb2eda77b848825ecdf5ef