[system-tests] Add pave test with old Zedboot
Test that paving using old Zedboot works with the current build.

This should prevent accidental changes that alter the image format.
Breaking changes to the image format are supported by submitting the
change to Zedboot to support it before changing how the image is
generated.

The test added here will be added to CQ in a follow up CL (which is in a
different repository).

Bug: 66500
Change-Id: I615d740e13b29df00e837cc3e7939a417a1c90ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480557
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Chris Suter <csuter@google.com>