[boot-shim] Mark the kernel item to be ignored when not removing it
The ZBI protocol requires that the data ZBI delivered by the boot
loader contain all the items after the bootable kernel item, and
allows it to contain other boot loader-supplied items.  But it
should not (still) contain the original kernel item.

The qemu boot shim was changed to avoid data-copying to get the
data ZBI aligned without the original kernel item, but this made
it violate the protocol by making the kernel item part of the
data ZBI.  This works fine for Zircon, but is confusing for other
cases like phys tests.  It's acceptable to permute the data ZBI
somewhat however, so a simple solution is to just change the item
type to DISCARD.

Change-Id: Ie1ad5b367a4de508362d42a5a378a30817beac82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485182
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>