[docs] Adds instructions about linking to in-tree code.
The style guide didn't have this piece of instruction, and it's
important because the static checks will fail without it.

Also, formatted the document to be 80 columns wide.  This is literally
what the document says the documentation style guide is, in the
section "Follow the 80 character limit".

Change-Id: I4c0f04201981e310a794eecd9a35229d3c9ecae8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480777
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>