[gidl] Handle rights parsing for GIDL
This adds support for parsing handle rights to the GIDL frontend for
handle defs (which specify handle inputs). Handle rights used for
validation are specified on the FIDL types.

Examples:
{ #0 = event(rights: execute + write) }
{ #0 = channel(rights: channel_default - signal) }

This CL redefined the handle rights constants because there aren't
handle rights constants that are accessible to GIDL in go (and the
existing go handle rights constants are a partial set). They are defined
as a switch over strings because it seemed simpler than defining
constants in one place and then having a similar list elsewhere.

Bug:41920
Change-Id: Id20ddbad656c8634d7c94c5ffe01e67e485bad9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484117
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>