[fidl][fidlconv] Support contextual built-in type resolution.
Previously, changes that required knowledge of the type underpinning a
particular identifier were impossible: if we're converting something
like "struct S { Foo a?; };", what does "Foo" mean?  We'd like to pursue
different conversions based on on the underlying type (ie, if "Foo" is a
struct, we'll need to box it, etc).

This change solves this problem by processing the file being converted
twice: once to generate a flat::Library, and a second time to generate
the raw AST for the ConvertingTreeVisitor to visit.  The flat::Library
representation is then passed to the ConvertingTreeVisitor, so that it
may use it to look up the underlying type when performing a
TypeConversion.

Bug: 68667
Test: fidldev test fidlc
Change-Id: Ib4c066f66ef63eea8cad8de20f9f4f28fcd891d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481159
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>