[fidlc] Emit obj_type for handles
Handles are mostly generalized, and introduced by `resource_declaration`
which specify an `obj_type`, itself an `enum` of a `uint32`. Exposing
the resolved value of an `obj_type` in the JSON IR to allow backends to
rely on this value, rather than duplicate the enum whose source of truth
is in a FIDL library (today, `library zx;` for Zircon handles).

In a next change, this new JSON IR value will be used to remove the
`ObjectTypeFromHandleSubtype` helper in fidlgen lib, along with
constants which currently duplicate the object type definition present
in the source FIDL library.

Bug: 64629
Bug: 45998
Test: fidlc tests and fidlc goldens
Change-Id: I8cd695cb429b45ef66fac671299560040fc81eaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477577
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Fuchsia-Auto-Submit: Pascal Perez <pascallouis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>