[paver]: Fix Unaligned Access.
This is cause for UB San failure. The problem is when accessing a
PartitionDescriptor/ExtentDescriptor, which is a serialized
struct(oacked) within a sequence of this packed struct, it is not
guaranteed that access is aligned.

Instead of reinterpreted pointer, memcpy to a aligned storage.

Bug: 57974 #done
Test: CQ
Change-Id: Iee46adeb3b68fb45cf31f273dc0c99a52ee95e3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484557
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Chris Suter <csuter@google.com>