[sdk][gn] Remove intermediate files after generating an archive.
This ensures that gn_sdk targets do not create new build outputs
when given the same build inputs. This fixes a bug where this target
would not successfully build on CI/CQ where Ninja fails if the build
is not "convergent".

gn_sdk has been reduced to a function: `f(in_archive) -> out_archive`
to hide the pieces of the GN implementation that are not yet working.

//sdk:core.modular now builds this new archive to ensure it continues
working on CI/CQ.

Test:
 * CQ: See 'core.x64-sdk-modular' tryjob.
 * Local:
    * fx set core.x64 --with //scripts/sdk/gn --args="build_sdk_archives=true"
    * fx build && fx build
    * Verify ninja says "no work to do" on the second build.

Bug: 69444
Change-Id: I22f0033183e721e286a9eccd2773cb91117ae600
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483718
Commit-Queue: Kendal Harland <kjharland@google.com>
Reviewed-by: Sean Cuff <seancuff@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>