[wlan][policy] Recovers deleted interfaces.
This change checks notifications of deleted interfaces to
ensure that the IfaceManager was the one that requested the
deletion.  If the IfaceManager did not instigate the
deletion, the IfaceManager will attempt to recreate the
interface.

This is an initial attempt at recovering in instances where
the driver takes corrective action by deleting interfaces.

Bug: 60089
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: I5407d45b6364aa622daeacb9bf7969a613537280
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460756
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>