[fidlgen] Remove uses of ObjectTypeFromHandleSubtype
As we move towards properly generalizing handles, we can rely on the
object type emitted by the JSON IR in fidlgen backends. By doing so,
there are less ways for things to get out-of-sync. As an example, this
improvement fixed the handling of two object types clock and exceptions.
These are reflected in the goldens of the affected fidlgen backends (Go,
Dart).

While it would be nice to remove ObjectTypeFromHandleSubtype altogether,
we cannot do so quite yet. In GIDL, we parse handles with a nice looking
syntax like `#0 = event(rights: execute + write )`. There must be a
place to map `event` to its meaning, and since some GIDL backends --
looking at you fuzzer_corpus -- rely on the underlying object type
value, we must correspondibngly have a mapping to object types.

Lastly, since the mapping of clock and exception was missing, this was
added so that the converter is more correct.

Bug: 64629
Bug: 45998
Change-Id: Id6162bb235b97236785fda0f2084a0a265198c2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477578
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Pascal Perez <pascallouis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>