[fxfs] Fixed volume creation.
Object stores need to be registered with store manager so that we can
find them when replaying the log. When we're replaying the log we need a
way to lazily open stores.

Volume information is stored in a new reserved object (0) that exists
for all object stores. This volume information points at the root
directory for the volume.

Change-Id: I13b055f9f0df8d95da977a09661ba9d2e0162986