[fs] Derive PagedVfs from ManagedVfs.
ManagedVfs isused on Fuchsia for Blobfs to track the FIDL connections.
So if Blobfs is going to use PagedVfs eventually it needs to be derived
from ManagedVfs.

This design makes sense because ManagedVfs is designed to be run on
Fuchsia with FIDL, and paging will only work in this situation.
Long-term the paging functionality should be moved to ManagedVfs.

This also simplifies the test because ManagedVfs implements the
bookkeeping needed.

Change-Id: I106d1f57e94f17095f3f5b1c1a13f792208ebf88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485582
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>