[sl4f] Check for null currentUrl from WebDriver
Sometimes a WebDriver session returns null as the current URL, which
throws an exception when passed to `Uri.parse`. I'm not certain of the
exact conditions required to get a null response, but it might be due to
the WebDriver session ending as the request for the URL is in flight.

Bug: 68077, b/180029795
Change-Id: Ibd6499ea42d768bab87174073202e733f9e7392d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485693
Fuchsia-Auto-Submit: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Isma Cortes <isma@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>