[setui] Use interior mutability in DeviceStorage
This change moves the cached values and stash proxy behind a mutex
inside of DeviceStorage. The get and write methods are also changed
to only need &self, which allows us to remove the need for a Mutex
in all clients of DeviceStorage.

This changes also removes some unneeded [cfg(test)] attributes in
use statements in tests.

Bug: 67371
Test: fx test -o setui_service_tests
Change-Id: Ib56ae47d0ab91ee35ae83f7f3ce9a7d0d29c43e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485438
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>