[codec_impl] Allow adapters to set profiles
The CodecImpl library uses a pair of threads internally: a shared_fidl
thread provided by the caller and an internal stream control thread.
Additionally, the DecryptorAdapter has an processing loop thread for
performing decryption operations. As a server library, CodecImpl is
largely unaware of performance or latency requirements, so we should
allow the Codec/DecryptorAdapter implementations set scheduler profiles.

Scheduler profiles can be acquired in a number of ways, but we should
allow CodecImpl users to use the least permissive approach, which is
fuchsia.media.ProfileProvider. That requires the thread handle to be
provided, so this CL exposes the stream control thread and the decryptor
processing thread to the Adapter implementations.

Bug: 67050
Change-Id: Ia0167942e2db7bc0a4027c6d64b9490aad3382cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485520
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>