[settings] Add camera watcher agent
Adds an agent to watch the fuchsia.camera3.Device api for changes in the
camera's mute state. The agent will pick up changes and forward them on
to the settings service to keep state in sync.

Bug: 68538
Test: fx test -o setui_service_tests

The boilerplate testing code in camera_watcher.rs and
camera_watcher_agent_tests.rs are mostly the same as in media_buttons.

Change-Id: If2deebf3b057c031a2d94bd4e701014d78784b06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483348
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: William Xiao <wxyz@google.com>