[zxdb] Temporarily disable DwarfSymbolFactory.InlinedMemberFunction
Temporarily disable this test to facilitate the Clang roll. This test is
failing as of 1d68e0a0756476ceee1e62a778021e339c631940 because it checks
for the "this" parameter in a subroutine which Clang not omits because
it only has `abstract_origin` as an attribute. Clang is allowed to omit
this because it has no meaningful attributes.

After the roll, we will come back and re-enable this test with a fix.

Multiply: host_x64/zxdb_tests: 1
Bug: 69724
Change-Id: I8345ad1c6a93483e351e3b8dcd8d58c22758cf2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485156
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>