[kazoo][go] Add fence between vdso{PC,SP}
In https://github.com/golang/go/issues/40823, it's hypothesized that
some test failures we observe are due to assumptions that, when `vdsoSP`
is set, `vdsoPC` is also valid. ARM may reorder stores (which is
possibly further exacerbated on multi-socket machines) to these fields;
insert a memory barrier between the production so that if `vdsoSP` is
observed to be set, `vdsoPC` is guaranteed to be observable.

This changes the syscall generation for Go, and relevant testing is
occurring there in the various multithreaded pprof tests.

Bug: 69510

Change-Id: I7ddfec5b5c5366f7ec4fa2342a5bfcbe772f2515
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485188
Reviewed-by: Scott Graham <scottmg@google.com>
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>