[wlan][policy] Determine WPA3 support in policy
Adds a check to determine whether the device supports WPA3 in the
policy layer. This is used to reject connect calls to WPA3 networks when
WPA3 is not supported, and to treat mixed WPA2/WPA3 as WPA2 or WPA3 based
on whether WPA3 is supported. Network selection is not effected because
the lower layer already factors in  WPA3 support when determining whether
a network is compatible.

Multiply: wlancfg-tests
Test: added unit tests:
	fx test wlancfg-tests
      Manually tested that if WPA3 is not supported the connect
      call returns as not supported and if WPA3 is supported the
      connect succeeds. On reboot if a WPA3 network is saved and WPA3
      is supported we will attempt to auto connect to the WPA3 network.
Bug: 67274
Change-Id: Id3665f637d2073850ecefee524983e4f08abb41c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460455
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>