[fint] Set --ninja-executable GN flag
For consistency with the current version of `fx set`. See
fxrev.dev/461355 for the full rationale of enabling it in `fx set`. It's
effectively a no-op on the infrastructure because infra always does a
clean build, and it doesn't significantly impact the runtime of `gn gen`
so there's no reason to not run it on infra as well as in fx.

Bug: 68465
Change-Id: I0ecddce0aa33d7c9eea6ff39533caf6004cf2334
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484899
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Boone <gboone@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>