[fs] Actually support typed channels in service connector
The ability to construct a |fs::Service| using a lambda that takes a
|fidl::ServerEnd| did not actually work in practice. |fidl::ServerEnd|
can be implicitly constructed from a channel today, but we are gating
that unsafe functionality behind an allow-list. It turned out that
passing a lambda taking a ServerEnd to |fs::Service| actually invoked
the constructor overload with the raw channel connector, due to the
implicit conversions. The templated constructor was never functioning,
because constructor template auto-deduction does not take into account
of implicit conversions (in this case, conversions into fit::function).
For example, calling a constructor taking a templated fit::function with
something that is not a fit::function (e.g. a lambda) will not trigger
template deduction.

Unfortunately, to be able to accurately express that semantics required
fairly invoked SFINAE. I couldn't yet think of a more concise solution
that would enable the typed channel API without ugly workarounds.

TEST: Because |fidl::ServerEnd| still supports implicit conversions from
zx::channel today, the constructor is tested in a different CL that
removes the implicit conversions for parts of the code base:
fxrev.dev/I74a349e08190df75abd21b3fb6286d84687dc553

Change-Id: I06a55d6dce70adef957d259edfbf1aaa64063e9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485902
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>