[kernel][efi] Ensure EFI functions are called with the Microsoft ABI.
EFI-provided functions are defined to use the Microsoft calling
convention ABI. For the majority of our code that calls into EFI
services (e.g. Gigaboot), the entire executable is compiled with the MS
ABI, and everything works fine.

But we have a few locations in the Zircon kernel (_not_ compiled with
the MS ABI) that call into these EFI functions. This is completely
undefined behaviour, and has the potential to cause arbitrary problems.
In practice, they cause the EFI calls to mostly fail, and cause some
minor register corruption that has thus far gone unnoticed. It does mean
that EFI crashlog support is unlikely to have worked, however.

This CL tags the EFI functions with the "ms_abi" attribute that
instructs the compiler to use the correct calling convention.

This is a partial rollback of I34ecb8698572a33c4fa945dc62c7af0a60f04405.

Bug: 69268
Change-Id: I26f5b8ed681114b1ef510f04c55233fd56e40e3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481825
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>