[kernel][pc] Ensure EFI support exists on known-good platforms.
EFI is supported on most (e.g. NUC), but not all (e.g. QEMU), of our
x86_64 platforms. While we currently have tests to ensure that EFI calls
can be made, because EFI is not universal, the tests only do so if EFI
is detected.

One risk is that we inadvertently break EFI support in a way such that
EFI support can no longer be detected, but all our tests continue to
pass.

This test adds a list of "known good" platforms that are expected to
have working EFI support. If EFI cannot be detected on such a platform,
we fail the test.

Bug: 69268
Change-Id: I58f2226caa30dc6826f871ba2fb33725919d6b2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483009
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>