[bt][hfp] Add GainControl to handle remote volume control.
The call manager can request changes to the remote peer's gain for
speaker or microphone. These requests should be passed along to the peer
when they are supported. They are represented as a stream of items that
can be polled from GainControl.

When supported, the peer can report current gain for speaker or
microphone as HF indicators. These reports should be passed along to the
call service.

Bug: fxbug.dev/64566
Bug: fxbug.dev/64498
Test: Added unit tests for GainControl

Change-Id: I8d2caa6b5cb82f1b5e61854a2b027bdbce6a55e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462140
Commit-Queue: Jeff Belgum <belgum@google.com>
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Matthew Kehrt <kehrt@google.com>