[kernel][lib/arch] Update LBR code
This change:
* Updates miscellany for style and clarity.
* Removes the dependency of lib/arch on fbl
* Rewrites all of the LBR tests. They were made overly convoluted in
aiming for maximal coverage of the iteration mechanics across the legacy
LBR formats, which ultimately has lead to poor RoI and readability. We
now exercise the logic that is directly a function of the legacy formats
in more direct unit tests and greatly simplify the scope of the iteration
tests. Further simplification is made in replacing the use of
hwreg::MockIo with arch::testing::FakeMsrIo.

Change-Id: I5b1b86d3bb9565f5b0f9bca49dc2ca2dff28ad16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477484
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>