[setui] Make DeviceStorage cache dynamic
This change removes the generic parameter from the DeviceStorage struct,
add generic parameters to the get and write methods, and stores an Any
in the cache in place of the concrete type.

This is in preparation for completely removing the single type storage
dependency on the controllers.

In order to minimize changes, PhantomData is being used in ClientProxy
types to limit the scope of type changes that are necessary. Those will
be updated in a future CL.

Bug: 67371
Tests: fx test -o setui_service_tests
Also verified that storage still works by updating state and also
verifying it persisted across reboots.

Change-Id: I069de055e2673a126fdac807bd3814eeade2b4bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485757
Commit-Queue: Paul Faria <paulfaria@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: William Xiao <wxyz@google.com>