[netstack] Add the endpoint before callback registration
Fix a remote race where the hangup event is delivered soon after
callback registration and before the endpoint gets added to the
endpoints map. This race can cause the callback code to attempt
to remove the endpoint from the map before it was added.

Bug: 35594
Change-Id: I89cb24c7e6bcab07f268d4f4aedf7e75b9f5819c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486777
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>