[fs] Add locking to the PagedVnode.
Adds an inititial set of locking and lock annotations to the PagedVnode
to allow multithreaded access. The pager functions on the PagedVfs are
changed to allow the vmo_ access to be locked.

There is still a problem shutting down. The start of a new tracking
scheme is added with the addition of a Vnode::WillShutdown() function.
I will follow this up with some code in the VFS that tracks the
associated Vnodes better.

Change-Id: Id1b88ae274a4b4667aba44aafac09b2e0f95cfb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487078
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>