[bt][a2dp] Fix peers racing discovery to streaming
When we attempt to start streaming, the discovery and capability
collection process can take long enough that the peer starts
their own connection process after we have started collecting.

Previously we would attempt to start streaming and fail in this case,
after this change we will not attempt to start.

This is a reland of 480219.

Bug: fxbug.dev/67408
Bug: b/178721567
Test: added streaming_start_configure_while_discovery
Change-Id: Ia648f3cb5ff1fd82dde165df21eb3e85e90f5377
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484861
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jeff Belgum <belgum@google.com>