[kernel][tracing] Add WaitQueue depth trace events.
Add a couple of trace probe events to WaitQueueCollection::Insert and
WaitQueueCollection::Remove which log the queue's pointer, and depth
after the operation.  These probes allow scripts to analyze traced to
determine what a typical N is for a wait queue, and how frequently
threads are added to or removed from a queue with a depth of N.

Because of the obvious potential performance impact, the tracing is
compile-time disabled by default, and can be enabled using the newly
introduced wait_queue_depth_tracing_enabled kernel build flag.

Change-Id: I7c0f4badf3db52553a38b54989ecf21a22167a50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485593
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>