[host-target-testing] Fix retry logic for artifact downloads.
The download() function would cleanup by removing the dst directory when
it encountered an error before retrying. This caused a problem when
using a srcs-file for blobs because the srcs-file was created as a temp
file in the dst directory so it got deleted before it could retry the
download. Another bug was that the blobs were being downloaded to the
`packages` destination directory which already contained some files, so
by deleting the entire directory on a retry, it removed previously
downloaded files in that directory. The fix downloads everything to
a temp directory first before moving it to the specified destination.

Bug: 69467
Change-Id: I2f43f3d2a4ef4bccef76002c5882ea25648d1768
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485218
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>