[blobfs] Fix BlobCompressor::Create and PrepareWrite to correctly return errors.
There was a bug in the current implementation of PrepareWrite,
wherein, even if BlobCompressor::Create fails, we always
return ZX_OK success. This is because, BlobCompressor::Create
returns a std::optional instead of also returning the error.
While this is good programming practice for C++, we ended
up introducing a bug as we always return ZX_OK even if
BlobCompressor::Create fails.
Ideally, BlobCompressor::Create should return the error
it encountered in addition to std::optional, but that is a much
more involved change, and I will do it as a followup CL.

Bug:58097

Testing: Added BlobPrepareWriteFailure to test if we are returning error
         properly.
Change-Id: I3f93a0ef064879a4f837a25b62e8aa6a9635d965
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486557
Commit-Queue: Manali Bhutiyani <manalib@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>