[wlan] Add WPA3 support to Authenticator
Now that Authenticator supports WPA3, a test for a bug in
Supplicant (fxbug.dev/69303) is added. The bug is fixed in this
change, bug full integration testing is postponed to
http://fxrev.dev/c/484547/ as it requires refactoring of hw-sim.

Added a _private field to fourway::Config to ensure a compiler error
will result if the fourway::Config::new() function is not used to
construct one.

Bug: 69303
Test: fx test wlan-rsn-tests

Change-Id: Icf70b3e8628b96a789a8e42a042d9962ae1f2012
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485019
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>