[fx-set] Add support for build dir configuration
Decide which build dir to use depending on the value of the
_FX_BUILD_DIR variable, which is set if the top-level `fx --dir` flag is
set, and the --auto-dir flag.

I'm planning to stop supporting the --build-dir flag in this new version
of fx. It's been deprecated in favor of `fx --dir` for nearly two years,
with a warning printed on every use, so I think it's very likely safe to
remove. Supporting it would also make the already-complicated
conditional logic even uglier.

Bug: 68465
Change-Id: I05b193b673b1fa11946a6bbfe25cddf443f0e044
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487275
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>