[a11y] Refactor GestureContext to include multiple pointers.
As we start to support more complex multi-finger gestures, we need to
pass multiple fingers' state to downstream users (e.g. magnifier). This
change modifies the GestureContext struct to maintain state for all
pointers in a gesture. Along the way, I also eliminated the GestureInfo
type, as it was redundant with the GestureContext and refactored some of
the state logic in existing recognizers to leverage the new
GestureContext.

BUG: fxbug.dev/69465
TEST: fx test a11y-lib-tests; fx test a11y-tests

Change-Id: Ide832defff7e0d28581c7fc3f2e69851e781c1fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482401
Commit-Queue: Alexander Brusher <abrusher@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>