[blobfs] fail early if target_compression_size is small/incorrect.
This change adds a fail-early mechanism in blobfs-streaming writes,
if target_compression_size is set and incorrect/cannot accomodate
the minimum compressed size needed by the compressor.

As part of this change, we also need to allocate write_info_ field
in the blob, sooner that PrepareWrite, as we need to allow setting
the compression hint before calling ftruncate on the blob.

Testing: Added FailEarlyTargetCompressionSizeSet test to test new behavior.

Bug: 58097

Change-Id: I32f939fb5d8bf2cbdcf01cf2a117cd1881ea6bc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483679
Commit-Queue: Manali Bhutiyani <manalib@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>