[gigaboot] improve disk_find_partition() API
disk_find_partition() was a bit weird in that it modified the passed-in
disk information to "shrink" the first and last block to just the
requested partition, so it no longer really represented the whole disk.
It also took in a partition name parameter, but only ever used it for
debug logging, so wasn't actually able to filter by name (or GUID).

This CL modifies the function to be a more general partition search
function, able to look for any combination of type/GUID/name. This will
be useful going forward to start adapting the logic to the new
partitioning scheme.

Bug: 68692
Change-Id: Ideac88dc3f2387ca39204afbc61bdbf6755527eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484382
Reviewed-by: Varun Sharma <vars@google.com>
Commit-Queue: David Pursell <dpursell@google.com>