[vfs] Move lock to Vnode.
All minfs and blobfs files already had a lock so this will not increase
the object size. Directories and other uses of the VFS will now get a
lock that is currently unused, but these uses are rare.

This lock will be used to guard more data in a future pass (I want to
track Vfs pointers in the Vnode).

Change the Vnode::inflight_transactions_ from an atomic to a regular
value guarded by the lock. This should be almost the same and is
conceptually simpler.

There should be no behavior change.

Change-Id: I485f0853e0773e93df427ef2a4e693c31a5a4b5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487400
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>