[kernel][lib/arch] Introduce CacheConsistencyContext
Our arm64 version of arch::PostSelfModificationCacheSync was wrong (a
misunderstanding on the part of this author). This change replaces that
abstraction with an `arch::CacheConsistencyContext()` that separates out
the per-cache-line operations from per-cache operations in maintaining
cache coherence between the data and instruction caches.

Bug: 67615
Change-Id: I08ffa1afb0e25926262377e0d21fed485f4b737f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477798
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>