[vfs] Add Vfs backpointer.
Add a new constructor to Vnode that takes a Vfs pointer. The 0-argument
constructor will be deleted in a followup.

Tracks live Vnodes in the Vfs object and notifies them when the Vfs is
going away. This is like a weak pointer but we have access to the
backpointers.

Rename Vnode::WillShutdown to WillDeleteVfs. There is already a
"shutdown" operation for the ManagedVfs and that's a different step. We
don't want to delete the backpointers during ManagedVfs' shutdown. The
new name makes this distinction more clear.

Bug: 70397

Change-Id: I8c72da7a6d0d2f335e9f99788ff73f0ff6d7fc08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487968
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>