[bsdsocket_test] Increase the write blocking iteration
Address a flake where the test setup fails to ensure that the socket
send buffer stays backed up and hence the socket write stays blocked,
during the test.

This failure is because fill_stream_send_buf can sometimes fail to fill
up the send buffers (zircon socket buffers + gVisor send buffer) before
returning. The tests rely on subsequent writes to block after calling
this function, and that assumption fails as the write sometimes does not
block and succeeds because there is still room in the socket buffer.

Stream socket write is asynchronous and the operation is spread across
FDIO/application, loopWrite and gVisor endpoint Write, and that brings
in lot of timing non-determinism. We have some workarounds in
fill_stream_send_buf for this behavior to perform failing write in a
loop N times before returning. Increasing the N value in this CL until
we have a better SO_SNDBUF implementation. This increases the test
running time by ~250ms on QEMU, 4 CPUs.

Bug: 67142
Change-Id: Ib656b3927c35516010d166872a6afcab6707e14c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487960
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>